refactors for building
